Added expansion of tabs with spaces, moved UART setup
Managed to add an apparently working routine to expand tabs so that things like the pin state display correctly. Not at all sure it's a good implementation, but it was what I could come up with late last night.
Had to reduce the font size in the text display to show all the data left to right. Not crazy about that, but I don't see a better solution. Should be good on a tablet... bit hard to read on a phone. 
Moved the UART setup dialog code to it's own routine ahead of setting up another special handler for the "D" command... then found out D locks up the interface. Will probably have to remove it and use "d" for Multimeter mode instead with the phone requesting single updates.