Add Parameter Entry Dialog
When a command needs parameters, we figure out what to ask for from the description. E.g. #KHz means "add an edit field in the dialog box with KHz as the prompt, and accept numbers only." So "Generate #KHz %Duty" will, when selected, popup a dialog asking for the KHz and Duty and after those values are entered, it will put them into the command string to be sent.