Parameter Entry with Options Lists
Add the ability to define and display a list of options for the parameter selection wizard. e.g. 1=one 2=two. Somewhat limited at this point as the text of the option must be a single word. e.g. 1=this one results in two entries in the spinner: "1=this" and "one". The text before the = (if present) is used as the output. 

Add parameter wizard entries for 
- I2C
- SPI
- 2 Wire
- 3 Wire

Sadly, the SPI doesn't work because the BusPirate (firmware 6.1) doesn't accept the mode entry command in one line. I do 
m3 1 1 1 1 1
And it drops right into uart mode but if I do
m5 1 1 1 1 1
It stops and asks me for the speed. Other modes (besides m3) also work fine. Just m5, SPI. 

Corrected bug where old parameters accumulate between uses of the wizard. Wasn't clearing a global array between uses.