Change console font size to match line lengths displayed.
The console has always been hard to read at 8 points, but that was needed because otherwise the very long lines from the pinstate command or help list would fall off the right edge. Most commands don't begin to fill the screen, however, so a larger point size was very usable for most functions. This update adds a calculation for font size based on the longest line currently being displayed. When long lines scroll off the top, the size does not update until the next response, but this is a minor issue in practice.